+++ lib/sb.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:switchboard/pages/aboutPage.dart';
 import 'package:switchboard/pages/accountPage.dart';
+import 'package:switchboard/pages/editAlter.dart';
 import 'package:switchboard/pages/loginPage.dart';
 import 'package:switchboard/pages/privacyPolicyPage.dart';
 import 'package:switchboard/pages/registerPage.dart';
@@ -21,6 +22,7 @@ class Switchboard extends StatelessWidget {
         "/privacy": (ctx) => PrivacyPolicyPage(),
         "/account": (ctx) => AccountPage(),
         "/account/settings": (ctx) => SettingsPage(),
+        "/editAlter": (ctx) => EditAlterPage(),
       },
     );
   }