+++ server/migrations/0014_ChangeAltersPrimaryKey.php
@@ -1,0 +1,8 @@
+<?php
+
+return function($conn) {
+    $conn->QUERY("ALTER TABLE `Alters` DROP PRIMARY KEY;");
+    $conn->query("ALTER TABLE `Alters` ADD PRIMARY KEY (`ID`);");
+}
+
+?>